fix(ci): linux clippy casts/await-lock, alsa dep, prettier format 157 files
- mouse.rs: drop redundant i32 casts (location returns i32); allow await_holding_lock in drag (explicit drop, intentional)
- capture.rs: drop redundant u32 casts (xcap width/height return u32)
- ci/release.yml: add libasound2-dev (alsa-sys)
- frontend: prettier --write all src (157 files, was never formatted)

+++ src-tauri/crates/desktop-api/src/vision/capture.rs
@@ -115,8 +115,8 @@ async fn capture_fullscreen_and_crop(hwnd: isize) -> Result<Frame> {
 
         let x = win.x().unwrap_or(0).max(0) as u32;
         let y = win.y().unwrap_or(0).max(0) as u32;
-        let w = win.width().unwrap_or(0) as u32;
-        let h = win.height().unwrap_or(0) as u32;
+        let w = win.width().unwrap_or(0);
+        let h = win.height().unwrap_or(0);
 
         frame
             .crop(x, y, w, h)
